fix: int32

+++ ledger/dto.go
@@ -3,7 +3,7 @@ package main
 type LedgerMessage struct {
 	OrderID   string `json:"order_id"`
 	UserID    string `json:"user_id"`
-	Amount    int64  `json:"amount"`
+	Amount    int32  `json:"amount"`
 	Operation string `json:"operation"`
 	Date      string `json:"date"`
 }
